# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TransactionInput {
  /** Transaction amount in USD */
  amount: number;
  /** Merchant category code */
  merchantCategory: string;
  /** Hour of day (0-23) */
  hour: number;
  /** Day of week (0=Mon, 6=Sun) */
  dayOfWeek: number;
  /** Distance from home in km */
  distanceFromHome: number;
  /** Whether card was physically present */
  cardPresent: boolean;
  /** Whether it was an online transaction */
  onlineTransaction: boolean;
  /** Number of transactions in last 24 hours */
  numTransactionsLast24h: number;
  /** Average transaction amount for this card */
  avgTransactionAmount: number;
}

export type PredictionResultRiskLevel =
  (typeof PredictionResultRiskLevel)[keyof typeof PredictionResultRiskLevel];

export const PredictionResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface PredictionResult {
  transactionId: string;
  /** Probability of fraud (0-1) */
  fraudProbability: number;
  isFraud: boolean;
  riskLevel: PredictionResultRiskLevel;
  riskFactors: string[];
  confidence: number;
}

export interface BatchTransactionInput {
  transactions: TransactionInput[];
}

export interface BatchPredictionResult {
  totalProcessed: number;
  fraudDetected: number;
  results: PredictionResult[];
  processingTimeMs: number;
}

export type TransactionRiskLevel =
  (typeof TransactionRiskLevel)[keyof typeof TransactionRiskLevel];

export const TransactionRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Transaction {
  id: string;
  amount: number;
  merchantCategory: string;
  merchantName: string;
  timestamp: string;
  fraudProbability: number;
  isFraud: boolean;
  riskLevel: TransactionRiskLevel;
  cardLast4: string;
  location: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
}

export interface InsightsSummary {
  totalTransactions: number;
  totalFraudDetected: number;
  fraudRate: number;
  totalAmountAnalyzed: number;
  totalFraudAmount: number;
  avgFraudProbability: number;
  highRiskTransactions: number;
  savedAmount: number;
}

export interface TrendPoint {
  date: string;
  total: number;
  fraudCount: number;
  fraudRate: number;
  amount: number;
}

export interface TrendsResponse {
  daily: TrendPoint[];
  weekly: TrendPoint[];
}

export interface FeatureImportance {
  feature: string;
  importance: number;
  description: string;
}

export interface FeatureImportanceResponse {
  features: FeatureImportance[];
}

export type ListTransactionsParams = {
  limit?: number;
  offset?: number;
  fraudOnly?: boolean;
};