# lib/api-zod/src/generated/types/batchPredictionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionResult } from "./predictionResult";

export interface BatchPredictionResult {
  totalProcessed: number;
  fraudDetected: number;
  results: PredictionResult[];
  processingTimeMs: number;
}