# lib/api-zod/src/generated/types/batchTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInput } from "./transactionInput";

export interface BatchTransactionInput {
  transactions: TransactionInput[];
}