# lib/api-zod/src/generated/types/insightsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface InsightsSummary {
  totalTransactions: number;
  totalFraudDetected: number;
  fraudRate: number;
  totalAmountAnalyzed: number;
  totalFraudAmount: number;
  avgFraudProbability: number;
  highRiskTransactions: number;
  savedAmount: number;
}