# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionResultRiskLevel } from "./predictionResultRiskLevel";

export interface PredictionResult {
  transactionId: string;
  /** Probability of fraud (0-1) */
  fraudProbability: number;
  isFraud: boolean;
  riskLevel: PredictionResultRiskLevel;
  riskFactors: string[];
  confidence: number;
}