# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionRiskLevel } from "./transactionRiskLevel";

export interface Transaction {
  id: string;
  amount: number;
  merchantCategory: string;
  merchantName: string;
  timestamp: Date;
  fraudProbability: number;
  isFraud: boolean;
  riskLevel: TransactionRiskLevel;
  cardLast4: string;
  location: string;
}