# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionInput {
  /** Transaction amount in USD */
  amount: number;
  /** Merchant category code */
  merchantCategory: string;
  /** Hour of day (0-23) */
  hour: number;
  /** Day of week (0=Mon, 6=Sun) */
  dayOfWeek: number;
  /** Distance from home in km */
  distanceFromHome: number;
  /** Whether card was physically present */
  cardPresent: boolean;
  /** Whether it was an online transaction */
  onlineTransaction: boolean;
  /** Number of transactions in last 24 hours */
  numTransactionsLast24h: number;
  /** Average transaction amount for this card */
  avgTransactionAmount: number;
}