# lib/api-zod/src/generated/types/trendsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from "./trendPoint";

export interface TrendsResponse {
  daily: TrendPoint[];
  weekly: TrendPoint[];
}